backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
jinja2
google-generativeai
python-dotenv
